refactor: refactor service into trace and log
