+++ rest/app.py
@@ -6,20 +6,9 @@ from slowapi import Limiter, _rate_limit_exceeded_handler
 from slowapi.errors import RateLimitExceeded
 from slowapi.util import get_remote_address
 
-try:
-    from rest.service.ee.aws_client import TraceRootAWSClient
-except ImportError:
-    from rest.service.aws_client import TraceRootAWSClient
-
-try:
-    from rest.service.ee.tencent_client import TraceRootTencentClient
-except ImportError:
-    from rest.service.tencent_client import TraceRootTencentClient
-
 from rest.routers.auth import router as auth_router
 from rest.routers.explore import ExploreRouter
 from rest.routers.integrate import IntegrateRouter
-from rest.service.jaeger_client import TraceRootJaegerClient
 
 try:
     from rest.routers.ee.verify import VerifyRouter
@@ -45,19 +34,11 @@ class App:
             _rate_limit_exceeded_handler,
         )
         self.local_mode = os.getenv("REST_LOCAL_MODE", "false").lower() == "true"
-        self.observe_provider = os.getenv("REST_OBSERVE_PROVIDER", "aws").lower()
 
         # Add CORS middleware
         self.add_middleware()
 
-        if self.local_mode:
-            client = TraceRootJaegerClient()
-        elif self.observe_provider == "tencent":
-            client = TraceRootTencentClient()
-        else:
-            client = TraceRootAWSClient()
-
-        self.explore_router = ExploreRouter(client, self.limiter)
+        self.explore_router = ExploreRouter(self.local_mode, self.limiter)
         self.app.include_router(
             self.explore_router.router,
             prefix="/v1/explore",
