+++ rest/service/log/__init__.py
@@ -1,0 +1,5 @@
+"""Log client implementations."""
+
+from rest.service.log.log_client import LogClient
+
+__all__ = ["LogClient"]
