+++ rest/service/log/aws_log_client.py
@@ -1,0 +1,40 @@
+from datetime import datetime
+
+from rest.config.log import TraceLogs
+from rest.service.log.log_client import LogClient
+
+
+class AWSLogClient(LogClient):
+    """Client for querying logs from AWS CloudWatch Logs.
+
+    Note: This is a stub implementation. Use the EE version for full functionality.
+    """
+
+    def __init__(self, aws_region: str | None = None):
+        """Initialize the AWS log client.
+
+        Args:
+            aws_region: AWS region
+        """
+        self.aws_region = aws_region or "us-west-2"
+
+    async def get_logs_by_trace_id(
+        self,
+        trace_id: str,
+        start_time: datetime,
+        end_time: datetime,
+        log_group_name: str | None = None,
+        log_search_term: str | None = None,
+    ) -> TraceLogs:
+        """Query logs by trace ID - stub implementation."""
+        return TraceLogs(logs=[])
+
+    async def get_trace_ids_from_logs(
+        self,
+        start_time: datetime,
+        end_time: datetime,
+        log_group_name: str,
+        search_term: str,
+    ) -> list[str]:
+        """Get trace IDs from logs - stub implementation."""
+        return []
