+++ rest/service/log/tencent_log_client.py
@@ -1,0 +1,37 @@
+from datetime import datetime
+
+from rest.config.log import TraceLogs
+from rest.service.log.log_client import LogClient
+
+
+class TencentLogClient(LogClient):
+    """Client for querying logs from Tencent CLS."""
+
+    def __init__(self, tencent_region: str | None = None):
+        """Initialize the Tencent log client.
+
+        Args:
+            tencent_region: Tencent Cloud region
+        """
+        self.tencent_region = tencent_region or "ap-hongkong"
+
+    async def get_logs_by_trace_id(
+        self,
+        trace_id: str,
+        start_time: datetime,
+        end_time: datetime,
+        log_group_name: str | None = None,
+        log_search_term: str | None = None,
+    ) -> TraceLogs:
+        """Query logs by trace ID."""
+        return TraceLogs(logs=[])
+
+    async def get_trace_ids_from_logs(
+        self,
+        start_time: datetime,
+        end_time: datetime,
+        log_group_name: str,
+        search_term: str,
+    ) -> list[str]:
+        """Get trace IDs from logs."""
+        return []
