+++ rest/service/trace/__init__.py
@@ -1,0 +1,5 @@
+"""Trace client implementations."""
+
+from rest.service.trace.trace_client import TraceClient
+
+__all__ = ["TraceClient"]
