+++ rest/service/trace/aws_trace_client.py
@@ -1,0 +1,46 @@
+from datetime import datetime
+
+from rest.config.trace import Span, Trace
+from rest.service.trace.trace_client import TraceClient
+
+
+class AWSTraceClient(TraceClient):
+    """Client for querying traces from AWS X-Ray.
+
+    Note: This is a stub implementation. Use the EE version for full functionality.
+    """
+
+    def __init__(self, aws_region: str | None = None):
+        """Initialize the AWS trace client.
+
+        Args:
+            aws_region: AWS region
+        """
+        self.aws_region = aws_region or "us-west-2"
+
+    async def get_trace_with_spans_by_ids(
+        self,
+        trace_ids: list[str],
+        categories: list[str] | None = None,
+        values: list[str] | None = None,
+        operations: list[str] | None = None,
+    ) -> dict[str,
+              list[Span]]:
+        """Get traces with spans by trace IDs - stub implementation."""
+        return {}
+
+    async def get_recent_traces(
+        self,
+        start_time: datetime,
+        end_time: datetime,
+        log_group_name: str,
+        service_name_values: list[str] | None = None,
+        service_name_operations: list[str] | None = None,
+        service_environment_values: list[str] | None = None,
+        service_environment_operations: list[str] | None = None,
+        categories: list[str] | None = None,
+        values: list[str] | None = None,
+        operations: list[str] | None = None,
+    ) -> list[Trace]:
+        """Get recent traces - stub implementation."""
+        return []
