+++ rest/service/trace/tencent_trace_client.py
@@ -1,0 +1,43 @@
+from datetime import datetime
+
+from rest.config.trace import Span, Trace
+from rest.service.trace.trace_client import TraceClient
+
+
+class TencentTraceClient(TraceClient):
+    """Client for querying traces from Tencent APM."""
+
+    def __init__(self, tencent_region: str | None = None):
+        """Initialize the Tencent trace client.
+
+        Args:
+            tencent_region: Tencent Cloud region
+        """
+        self.tencent_region = tencent_region or "ap-hongkong"
+
+    async def get_trace_with_spans_by_ids(
+        self,
+        trace_ids: list[str],
+        categories: list[str] | None = None,
+        values: list[str] | None = None,
+        operations: list[str] | None = None,
+    ) -> dict[str,
+              list[Span]]:
+        """Get traces with spans by trace IDs."""
+        return {}
+
+    async def get_recent_traces(
+        self,
+        start_time: datetime,
+        end_time: datetime,
+        log_group_name: str,
+        service_name_values: list[str] | None = None,
+        service_name_operations: list[str] | None = None,
+        service_environment_values: list[str] | None = None,
+        service_environment_operations: list[str] | None = None,
+        categories: list[str] | None = None,
+        values: list[str] | None = None,
+        operations: list[str] | None = None,
+    ) -> list[Trace]:
+        """Get recent traces."""
+        return []
