+++ rest/service/trace/trace_client.py
@@ -1,56 +1,13 @@
 from abc import ABC, abstractmethod
 from datetime import datetime
 
-from rest.config.log import TraceLogs
 from rest.config.trace import Span, Trace
 
 
-class TraceRootClient(ABC):
-    """Abstract base class for TraceRoot clients that query logs and traces
+class TraceClient(ABC):
+    """Abstract base class for trace clients that query traces
     from different observability backends.
     """
-
-    @abstractmethod
-    async def get_logs_by_trace_id(
-        self,
-        trace_id: str,
-        start_time: datetime,
-        end_time: datetime,
-        log_group_name: str | None = None,
-        log_search_term: str | None = None,
-    ) -> TraceLogs:
-        """Query logs by trace ID.
-
-        Args:
-            trace_id: Trace identifier
-            start_time: Start time of the trace
-            end_time: End time of the trace
-            log_group_name: Log group name (provider-specific)
-            log_search_term: Additional search term to filter logs
-
-        Returns:
-            TraceLogs: Trace logs for the given trace ID
-        """
-
-    @abstractmethod
-    async def get_trace_ids_from_logs(
-        self,
-        start_time: datetime,
-        end_time: datetime,
-        log_group_name: str,
-        search_term: str,
-    ) -> list[str]:
-        """Get trace IDs from logs.
-
-        Args:
-            start_time: Start time for log query
-            end_time: End time for log query
-            log_group_name: Log group name
-            search_term: Search term to look for in log messages
-
-        Returns:
-            List of unique trace IDs found in logs containing the search term
-        """
 
     @abstractmethod
     async def get_trace_with_spans_by_ids(
